scripts/requirements.txt
algoliasearch==1.5.8
clarifai==0.2
httplib2==0.9.1
python-instagram==1.3.1
simplejson==3.7.3
six==1.9.0
urllib3==1.10.4
wsgiref==0.1.2
